+++ CHANGELOG.md
@@ -14,7 +14,7 @@ DEPS_CLJ_TOOLS_VERSION=1.11.1.1165 bb clojure
 
 - [#104](https://github.com/borkdude/deps.clj/issues/104): print repl-aliases in -Sdescribe
 - Make installation of tools jar more robust using transaction file
-- #106: System deps.edn should be extracted on tools install
+- [#106](https://github.com/borkdude/deps.clj/issues/106): System deps.edn should be extracted on tools install
 
 ## 1.11.1.1347